% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{simulate_phi_hebart}
\alias{simulate_phi_hebart}
\title{Simulate mu_js}
\usage{
simulate_phi_hebart(tree, R, groups, tau, sigma_phi, M, num_trees)
}
\arguments{
\item{tree}{The current tree}

\item{R}{The corresponding residuals for the tree}

\item{groups}{The groups specification}

\item{tau}{The current value of tau}

\item{M}{The group allocation matrix}

\item{num_trees}{The  number of trees}

\item{tau_phi}{The current value of tau_phi}
}
\description{
Simulates mu_j for each group
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
